Refactor admin user API routes
